Revert "Renaming qconsolidate like stuff into oq_consolidate like"
This reverts commit 2bba8812c88ad3293d62f15bb50864d7dfd27341.
