+++ consolidatethread.py
