+++ icons/qconsolidate.png
