+++ qconsolidate.pro
@@ -1,0 +1,5 @@
+SOURCES = __init__.py \
+          qconsolidate.py \
+          qconsolidatedialog.py \
+          aboutdialog.py \
+          consolidatethread.py
