+++ resources.qrc
@@ -1,6 +1,6 @@
 <RCC>
   <qresource>
-    <file>icons/oq_consolidate.png</file>
+    <file>icons/qconsolidate.png</file>
     <file>icons/about.png</file>
   </qresource>
 </RCC>
