+++ resources_rc.py
@@ -208,11 +208,10 @@ qt_resource_name = b"\
 \x00\x6f\xa6\x53\
 \x00\x69\
 \x00\x63\x00\x6f\x00\x6e\x00\x73\
-\x00\x12\
-\x0e\x81\x61\xa7\
-\x00\x6f\
-\x00\x71\x00\x5f\x00\x63\x00\x6f\x00\x6e\x00\x73\x00\x6f\x00\x6c\x00\x69\x00\x64\x00\x61\x00\x74\x00\x65\x00\x2e\x00\x70\x00\x6e\
-\x00\x67\
+\x00\x10\
+\x04\x01\xe1\x87\
+\x00\x71\
+\x00\x63\x00\x6f\x00\x6e\x00\x73\x00\x6f\x00\x6c\x00\x69\x00\x64\x00\x61\x00\x74\x00\x65\x00\x2e\x00\x70\x00\x6e\x00\x67\
 \x00\x09\
 \x06\xc7\x98\x67\
 \x00\x61\
@@ -222,8 +221,8 @@ qt_resource_name = b"\
 qt_resource_struct = b"\
 \x00\x00\x00\x00\x00\x02\x00\x00\x00\x01\x00\x00\x00\x01\
 \x00\x00\x00\x00\x00\x02\x00\x00\x00\x02\x00\x00\x00\x02\
-\x00\x00\x00\x3a\x00\x00\x00\x00\x00\x01\x00\x00\x03\x00\
 \x00\x00\x00\x10\x00\x00\x00\x00\x00\x01\x00\x00\x00\x00\
+\x00\x00\x00\x36\x00\x00\x00\x00\x00\x01\x00\x00\x03\x00\
 "
 
 def qInitResources():
